requirements.txt
google-genai
requests
beautifulsoup4
google-cloud-service-usage
google-auth
google-api-python-client
python-dotenv